feat: separate tests for executing and formatting with new functional options that work for both test types
