+++ task_test.go
@@ -14,6 +14,7 @@ import (
 	"path/filepath"
 	"regexp"
 	"runtime"
+	"slices"
 	"sort"
 	"strings"
 	"sync"
@@ -32,146 +33,213 @@ import (
 	"github.com/go-task/task/v3/taskfile/ast"
 )
 
+func init() {
+	_ = os.Setenv("NO_COLOR", "1")
+}
+
 type (
-	// A PostProcessFn is a function that can be applied to the output of a test
-	// fixture before the file is written.
-	PostProcessFn func(*testing.T, []byte) []byte
-	// A TaskTestOption is a function that configures a [TaskTest].
-	TaskTestOption func(*TaskTest)
-	// A TaskTest is a test wrapper around a [task.Executor] to make it easy to
-	// write tests for tasks. See [NewTaskTest] for information on creating and
-	// running TaskTests. These tests use fixture files to assert whether the
-	// result of a task is correct. If Task's behavior has been changed, the
-	// fixture files can be updated by running `task gen:fixtures`.
+	TestOption interface {
+		ExecutorTestOption
+		FormatterTestOption
+	}
 	TaskTest struct {
-		name            string
-		task            string
-		vars            map[string]any
-		input           string
-		executorOpts    []task.ExecutorOption
-		experiments     map[*experiments.Experiment]int
-		postProcessFns  []PostProcessFn
-		wantSetupError  bool
-		wantRunError    bool
-		wantStatusError bool
+		name           string
+		experiments    map[*experiments.Experiment]int
+		postProcessFns []PostProcessFn
 	}
 )
 
-// NewTaskTest sets up a new [task.Executor] with the given options and runs a
-// task with the given [TaskTestOption]s. The output of the task is written to a
-// set of fixture files depending on the configuration of the test.
-func NewTaskTest(t *testing.T, opts ...TaskTestOption) {
+// goldenFileName makes the file path for fixture files safe for all well-known
+// operating systems. Windows in particular has a lot of restrictions the
+// characters that can be used in file paths.
+func goldenFileName(t *testing.T) string {
 	t.Helper()
-	tt := &TaskTest{
-		task:        "default",
-		vars:        map[string]any{},
-		experiments: map[*experiments.Experiment]int{},
+	name := t.Name()
+	for _, c := range []string{` `, `<`, `>`, `:`, `"`, `/`, `\`, `|`, `?`, `*`} {
+		name = strings.ReplaceAll(name, c, "-")
 	}
-	// Apply the functional options
-	for _, opt := range opts {
-		opt(tt)
+	return name
+}
+
+// writeFixture writes a fixture file for the test. The fixture file is created
+// using the [goldie.Goldie] package. The fixture file is created with the
+// output of the task, after any post-process functions have been applied.
+func (tt *TaskTest) writeFixture(
+	t *testing.T,
+	g *goldie.Goldie,
+	goldenFileSuffix string,
+	b []byte,
+) {
+	t.Helper()
+	// Apply any post-process functions
+	for _, fn := range tt.postProcessFns {
+		b = fn(t, b)
 	}
-	// Enable any experiments that have been set
-	for x, v := range tt.experiments {
-		prev := *x
-		*x = experiments.Experiment{
-			Name:          prev.Name,
-			AllowedValues: []int{v},
-			Value:         v,
-		}
-		t.Cleanup(func() {
-			*x = prev
-		})
+	// Write the fixture file
+	goldenFileName := goldenFileName(t)
+	if goldenFileSuffix != "" {
+		goldenFileName += "-" + goldenFileSuffix
 	}
-	tt.run(t)
+	g.Assert(t, goldenFileName, b)
+}
+
+// writeFixtureBuffer is a wrapper for writing the main output of the task to a
+// fixture file.
+func (tt *TaskTest) writeFixtureBuffer(
+	t *testing.T,
+	g *goldie.Goldie,
+	buff bytes.Buffer,
+) {
+	t.Helper()
+	tt.writeFixture(t, g, "", buff.Bytes())
+}
+
+// writeFixtureErrSetup is a wrapper for writing the output of an error during
+// the setup phase of the task to a fixture file.
+func (tt *TaskTest) writeFixtureErrSetup(
+	t *testing.T,
+	g *goldie.Goldie,
+	err error,
+) {
+	t.Helper()
+	tt.writeFixture(t, g, "err-setup", []byte(err.Error()))
 }
 
 // Functional options
 
 // WithName gives the test fixture output a name. This should be used when
-// running multiple [TaskTest]s in a single test function.
-func WithName(name string) TaskTestOption {
-	return func(tt *TaskTest) {
-		tt.name = name
-	}
+// running multiple tests in a single test function.
+func WithName(name string) TestOption {
+	return &nameTestOption{name: name}
+}
+
+type nameTestOption struct {
+	name string
+}
+
+func (opt *nameTestOption) applyToExecutorTest(t *ExecutorTest) {
+	t.name = opt.name
+}
+
+func (opt *nameTestOption) applyToFormatterTest(t *FormatterTest) {
+	t.name = opt.name
 }
 
 // WithTask sets the name of the task to run. This should be used when the task
 // to run is not the default task.
-func WithTask(task string) TaskTestOption {
-	return func(tt *TaskTest) {
-		tt.task = task
-	}
+func WithTask(task string) TestOption {
+	return &taskTestOption{task: task}
+}
+
+type taskTestOption struct {
+	task string
+}
+
+func (opt *taskTestOption) applyToExecutorTest(t *ExecutorTest) {
+	t.task = opt.task
+}
+
+func (opt *taskTestOption) applyToFormatterTest(t *FormatterTest) {
+	t.task = opt.task
 }
 
 // WithVar sets a variable to be passed to the task. This can be called multiple
 // times to set more than one variable.
-func WithVar(key string, value any) TaskTestOption {
-	return func(tt *TaskTest) {
-		tt.vars[key] = value
-	}
+func WithVar(key string, value any) TestOption {
+	return &varTestOption{key: key, value: value}
 }
 
-// WithInput tells the test to create a reader with the given input. This can be
-// used to simulate user input when a task requires it.
-func WithInput(input string) TaskTestOption {
-	return func(tt *TaskTest) {
-		tt.input = input
-	}
+type varTestOption struct {
+	key   string
+	value any
+}
+
+func (opt *varTestOption) applyToExecutorTest(t *ExecutorTest) {
+	t.vars[opt.key] = opt.value
+}
+
+func (opt *varTestOption) applyToFormatterTest(t *FormatterTest) {
+	t.vars[opt.key] = opt.value
 }
 
 // WithExecutorOptions sets the [task.ExecutorOption]s to be used when creating
-// [task.Executor].
-func WithExecutorOptions(executorOpts ...task.ExecutorOption) TaskTestOption {
-	return func(tt *TaskTest) {
-		tt.executorOpts = append(tt.executorOpts, executorOpts...)
-	}
+// a [task.Executor].
+func WithExecutorOptions(executorOpts ...task.ExecutorOption) TestOption {
+	return &executorOptionsTestOption{executorOpts: executorOpts}
+}
+
+type executorOptionsTestOption struct {
+	executorOpts []task.ExecutorOption
+}
+
+func (opt *executorOptionsTestOption) applyToExecutorTest(t *ExecutorTest) {
+	t.executorOpts = slices.Concat(t.executorOpts, opt.executorOpts)
+}
+
+func (opt *executorOptionsTestOption) applyToFormatterTest(t *FormatterTest) {
+	t.executorOpts = slices.Concat(t.executorOpts, opt.executorOpts)
 }
 
 // WithExperiment sets an experiment to be enabled for the test. This can be
 // called multiple times to enable more than one experiment.
-func WithExperiment(experiment *experiments.Experiment, value int) TaskTestOption {
-	return func(tt *TaskTest) {
-		tt.experiments[experiment] = value
-	}
+func WithExperiment(experiment *experiments.Experiment, value int) TestOption {
+	return &experimentTestOption{experiment: experiment, value: value}
+}
+
+type experimentTestOption struct {
+	experiment *experiments.Experiment
+	value      int
+}
+
+func (opt *experimentTestOption) applyToExecutorTest(t *ExecutorTest) {
+	t.experiments[opt.experiment] = opt.value
+}
+
+func (opt *experimentTestOption) applyToFormatterTest(t *FormatterTest) {
+	t.experiments[opt.experiment] = opt.value
 }
 
 // WithPostProcessFn adds a [PostProcessFn] function to the test. Post-process
 // functions are run on the output of the task before a fixture is created. This
 // can be used to remove absolute paths, sort lines, etc. This can be called
 // multiple times to add more than one post-process function.
-func WithPostProcessFn(fn PostProcessFn) TaskTestOption {
-	return func(tt *TaskTest) {
-		tt.postProcessFns = append(tt.postProcessFns, fn)
-	}
+func WithPostProcessFn(fn PostProcessFn) TestOption {
+	return &postProcessFnTestOption{fn: fn}
+}
+
+type postProcessFnTestOption struct {
+	fn PostProcessFn
+}
+
+func (opt *postProcessFnTestOption) applyToExecutorTest(t *ExecutorTest) {
+	t.postProcessFns = append(t.postProcessFns, opt.fn)
+}
+
+func (opt *postProcessFnTestOption) applyToFormatterTest(t *FormatterTest) {
+	t.postProcessFns = append(t.postProcessFns, opt.fn)
 }
 
 // WithSetupError sets the test to expect an error during the setup phase of the
 // task execution. A fixture will be created with the output of any errors.
-func WithSetupError() TaskTestOption {
-	return func(tt *TaskTest) {
-		tt.wantSetupError = true
-	}
+func WithSetupError() TestOption {
+	return &setupErrorTestOption{}
 }
 
-// WithRunError sets the test to expect an error during the run phase of the
-// task execution. A fixture will be created with the output of any errors.
-func WithRunError() TaskTestOption {
-	return func(tt *TaskTest) {
-		tt.wantRunError = true
-	}
+type setupErrorTestOption struct{}
+
+func (opt *setupErrorTestOption) applyToExecutorTest(t *ExecutorTest) {
+	t.wantSetupError = true
 }
 
-// WithStatusError tells the test to make an additional call to
-// [task.Executor.Status] after the task has been run. A fixture will be created
-// with the output of any errors.
-func WithStatusError() TaskTestOption {
-	return func(tt *TaskTest) {
-		tt.wantStatusError = true
-	}
+func (opt *setupErrorTestOption) applyToFormatterTest(t *FormatterTest) {
+	t.wantSetupError = true
 }
 
-// Post-process functions
+// Post-processing
+
+// A PostProcessFn is a function that can be applied to the output of a test
+// fixture before the file is written.
+type PostProcessFn func(*testing.T, []byte) []byte
 
 // PPRemoveAbsolutePaths removes any absolute paths from the output of the task.
 // This is useful when the task output contains paths that are can be different
@@ -192,165 +260,6 @@ func PPSortedLines(t *testing.T, b []byte) []byte {
 	lines := strings.Split(strings.TrimSpace(string(b)), "\n")
 	sort.Strings(lines)
 	return []byte(strings.Join(lines, "\n") + "\n")
-}
-
-// Helpers
-
-// goldenFileName makes the file path for fixture files safe for all well-known
-// operating systems. Windows in particular has a lot of restrictions the
-// characters that can be used in file paths.
-func goldenFileName(t *testing.T) string {
-	t.Helper()
-	name := t.Name()
-	for _, c := range []string{` `, `<`, `>`, `:`, `"`, `/`, `\`, `|`, `?`, `*`} {
-		name = strings.ReplaceAll(name, c, "-")
-	}
-	return name
-}
-
-// writeFixture writes a fixture file for the test. The fixture file is created
-// using the [goldie.Goldie] package. The fixture file is created with the
-// output of the task, after any post-process functions have been applied.
-func (tt *TaskTest) writeFixture(
-	t *testing.T,
-	g *goldie.Goldie,
-	goldenFileName string,
-	b []byte,
-) {
-	t.Helper()
-	// Apply any post-process functions
-	for _, fn := range tt.postProcessFns {
-		b = fn(t, b)
-	}
-	// Write the fixture file
-	g.Assert(t, goldenFileName, b)
-}
-
-// writeFixtureBuffer ir a wrapper for writing the main output of the task to a
-// fixture file.
-func (tt *TaskTest) writeFixtureBuffer(
-	t *testing.T,
-	g *goldie.Goldie,
-	buff bytes.Buffer,
-) {
-	t.Helper()
-	tt.writeFixture(t, g, goldenFileName(t), buff.Bytes())
-}
-
-// writeFixtureErrSetup is a wrapper for writing the output of an error during
-// the setup phase of the task to a fixture file.
-func (tt *TaskTest) writeFixtureErrSetup(
-	t *testing.T,
-	g *goldie.Goldie,
-	err error,
-) {
-	t.Helper()
-	tt.writeFixture(t, g, fmt.Sprintf("%s-err-setup", goldenFileName(t)), []byte(err.Error()))
-}
-
-// writeFixtureErrRun is a wrapper for writing the output of an error during the
-// run phase of the task to a fixture file.
-func (tt *TaskTest) writeFixtureErrRun(
-	t *testing.T,
-	g *goldie.Goldie,
-	err error,
-) {
-	t.Helper()
-	tt.writeFixture(t, g, fmt.Sprintf("%s-err-run", goldenFileName(t)), []byte(err.Error()))
-}
-
-// writeFixtureStatus is a wrapper for writing the output of an error when
-// making an additional call to [task.Executor.Status] to a fixture file.
-func (tt *TaskTest) writeFixtureStatus(
-	t *testing.T,
-	g *goldie.Goldie,
-	status string,
-) {
-	t.Helper()
-	tt.writeFixture(t, g, fmt.Sprintf("%s-err-status", goldenFileName(t)), []byte(status))
-}
-
-// run is the main function for running the test. It sets up the task executor,
-// runs the task, and writes the output to a fixture file.
-func (tt *TaskTest) run(t *testing.T) {
-	t.Helper()
-	f := func(t *testing.T) {
-		t.Helper()
-		var buf bytes.Buffer
-
-		opts := append(
-			tt.executorOpts,
-			task.ExecutorWithStdout(&buf),
-			task.ExecutorWithStderr(&buf),
-		)
-
-		// If the test has input, create a reader for it and add it to the
-		// executor options
-		if tt.input != "" {
-			var reader bytes.Buffer
-			reader.WriteString(tt.input)
-			opts = append(opts, task.ExecutorWithStdin(&reader))
-		}
-
-		// Set up the task executor
-		e := task.NewExecutor(opts...)
-
-		// Create a golden fixture file for the output
-		g := goldie.New(t,
-			goldie.WithFixtureDir(filepath.Join(e.Dir, "testdata")),
-		)
-
-		// Call setup and check for errors
-		if err := e.Setup(); tt.wantSetupError {
-			require.Error(t, err)
-			tt.writeFixtureErrSetup(t, g, err)
-			tt.writeFixtureBuffer(t, g, buf)
-			return
-		} else {
-			require.NoError(t, err)
-		}
-
-		// Create the task call
-		vars := ast.NewVars()
-		for key, value := range tt.vars {
-			vars.Set(key, ast.Var{Value: value})
-		}
-		call := &task.Call{
-			Task: tt.task,
-			Vars: vars,
-		}
-
-		// Run the task and check for errors
-		ctx := context.Background()
-		if err := e.Run(ctx, call); tt.wantRunError {
-			require.Error(t, err)
-			tt.writeFixtureErrRun(t, g, err)
-			tt.writeFixtureBuffer(t, g, buf)
-			return
-		} else {
-			require.NoError(t, err)
-		}
-
-		// If the status flag is set, run the status check
-		if tt.wantStatusError {
-			if err := e.Status(ctx, call); err != nil {
-				tt.writeFixtureStatus(t, g, err.Error())
-			}
-		}
-
-		tt.writeFixtureBuffer(t, g, buf)
-	}
-
-	// Run the test (with a name if it has one)
-	if tt.name != "" {
-		t.Run(tt.name, f)
-	} else {
-		f(t)
-	}
-}
-
-func init() {
-	_ = os.Setenv("NO_COLOR", "1")
 }
 
 // SyncBuffer is a threadsafe buffer for testing.
@@ -414,391 +323,6 @@ func (fct fileContentTest) Run(t *testing.T) {
 			assert.Equal(t, expectContent, s, "unexpected file content in %s", path)
 		})
 	}
-}
-
-func TestEmptyTask(t *testing.T) {
-	t.Parallel()
-	NewTaskTest(t,
-		WithExecutorOptions(
-			task.ExecutorWithDir("testdata/empty_task"),
-		),
-	)
-}
-
-func TestEmptyTaskfile(t *testing.T) {
-	t.Parallel()
-	NewTaskTest(t,
-		WithExecutorOptions(
-			task.ExecutorWithDir("testdata/empty_taskfile"),
-		),
-		WithSetupError(),
-		WithPostProcessFn(PPRemoveAbsolutePaths),
-	)
-}
-
-func TestEnv(t *testing.T) {
-	t.Setenv("QUX", "from_os")
-	NewTaskTest(t,
-		WithName("env precedence disabled"),
-		WithExecutorOptions(
-			task.ExecutorWithDir("testdata/env"),
-			task.ExecutorWithSilent(true),
-		),
-	)
-	NewTaskTest(t,
-		WithName("env precedence enabled"),
-		WithExecutorOptions(
-			task.ExecutorWithDir("testdata/env"),
-			task.ExecutorWithSilent(true),
-		),
-		WithExperiment(&experiments.EnvPrecedence, 1),
-	)
-}
-
-func TestVars(t *testing.T) {
-	t.Parallel()
-	NewTaskTest(t,
-		WithExecutorOptions(
-			task.ExecutorWithDir("testdata/vars"),
-			task.ExecutorWithSilent(true),
-		),
-	)
-}
-
-func TestRequires(t *testing.T) {
-	t.Parallel()
-	NewTaskTest(t,
-		WithName("required var missing"),
-		WithExecutorOptions(
-			task.ExecutorWithDir("testdata/requires"),
-		),
-		WithTask("missing-var"),
-		WithRunError(),
-	)
-	NewTaskTest(t,
-		WithName("required var ok"),
-		WithExecutorOptions(
-			task.ExecutorWithDir("testdata/requires"),
-		),
-		WithTask("missing-var"),
-		WithVar("FOO", "bar"),
-	)
-	NewTaskTest(t,
-		WithName("fails validation"),
-		WithExecutorOptions(
-			task.ExecutorWithDir("testdata/requires"),
-		),
-		WithTask("validation-var"),
-		WithVar("ENV", "dev"),
-		WithVar("FOO", "bar"),
-		WithRunError(),
-	)
-	NewTaskTest(t,
-		WithName("passes validation"),
-		WithExecutorOptions(
-			task.ExecutorWithDir("testdata/requires"),
-		),
-		WithTask("validation-var"),
-		WithVar("FOO", "one"),
-		WithVar("ENV", "dev"),
-	)
-	NewTaskTest(t,
-		WithName("required var missing + fails validation"),
-		WithExecutorOptions(
-			task.ExecutorWithDir("testdata/requires"),
-		),
-		WithTask("validation-var"),
-		WithRunError(),
-	)
-	NewTaskTest(t,
-		WithName("required var missing + fails validation"),
-		WithExecutorOptions(
-			task.ExecutorWithDir("testdata/requires"),
-		),
-		WithTask("validation-var-dynamic"),
-		WithVar("FOO", "one"),
-		WithVar("ENV", "dev"),
-	)
-	NewTaskTest(t,
-		WithName("require before compile"),
-		WithExecutorOptions(
-			task.ExecutorWithDir("testdata/requires"),
-		),
-		WithTask("require-before-compile"),
-		WithRunError(),
-	)
-	NewTaskTest(t,
-		WithName("var defined in task"),
-		WithExecutorOptions(
-			task.ExecutorWithDir("testdata/requires"),
-		),
-		WithTask("var-defined-in-task"),
-	)
-}
-
-// TODO: mock fs
-func TestSpecialVars(t *testing.T) {
-	t.Parallel()
-
-	const dir = "testdata/special_vars"
-	const subdir = "testdata/special_vars/subdir"
-	toAbs := func(rel string) string {
-		abs, err := filepath.Abs(rel)
-		assert.NoError(t, err)
-		return abs
-	}
-
-	tests := []struct {
-		target   string
-		expected string
-	}{
-		// Root
-		{target: "print-task", expected: "print-task"},
-		{target: "print-root-dir", expected: toAbs(dir)},
-		{target: "print-taskfile", expected: toAbs(dir) + "/Taskfile.yml"},
-		{target: "print-taskfile-dir", expected: toAbs(dir)},
-		{target: "print-task-version", expected: "unknown"},
-		{target: "print-task-dir", expected: toAbs(dir) + "/foo"},
-		// Included
-		{target: "included:print-task", expected: "included:print-task"},
-		{target: "included:print-root-dir", expected: toAbs(dir)},
-		{target: "included:print-taskfile", expected: toAbs(dir) + "/included/Taskfile.yml"},
-		{target: "included:print-taskfile-dir", expected: toAbs(dir) + "/included"},
-		{target: "included:print-task-version", expected: "unknown"},
-	}
-
-	for _, dir := range []string{dir, subdir} {
-		for _, test := range tests {
-			NewTaskTest(t,
-				WithName(fmt.Sprintf("%s-%s", dir, test.target)),
-				WithExecutorOptions(
-					task.ExecutorWithDir(dir),
-					task.ExecutorWithSilent(true),
-					task.ExecutorWithVersionCheck(true),
-				),
-				WithTask(test.target),
-				WithPostProcessFn(PPRemoveAbsolutePaths),
-			)
-		}
-	}
-}
-
-func TestConcurrency(t *testing.T) {
-	t.Parallel()
-	NewTaskTest(t,
-		WithExecutorOptions(
-			task.ExecutorWithDir("testdata/concurrency"),
-			task.ExecutorWithConcurrency(1),
-		),
-		WithPostProcessFn(PPSortedLines),
-	)
-}
-
-func TestParams(t *testing.T) {
-	t.Parallel()
-	NewTaskTest(t,
-		WithExecutorOptions(
-			task.ExecutorWithDir("testdata/params"),
-			task.ExecutorWithSilent(true),
-		),
-		WithPostProcessFn(PPSortedLines),
-	)
-}
-
-func TestDeps(t *testing.T) {
-	t.Parallel()
-	NewTaskTest(t,
-		WithExecutorOptions(
-			task.ExecutorWithDir("testdata/deps"),
-			task.ExecutorWithSilent(true),
-		),
-		WithPostProcessFn(PPSortedLines),
-	)
-}
-
-// TODO: mock fs
-func TestStatus(t *testing.T) {
-	t.Parallel()
-
-	const dir = "testdata/status"
-
-	files := []string{
-		"foo.txt",
-		"bar.txt",
-		"baz.txt",
-	}
-
-	for _, f := range files {
-		path := filepathext.SmartJoin(dir, f)
-		_ = os.Remove(path)
-		if _, err := os.Stat(path); err == nil {
-			t.Errorf("File should not exist: %v", err)
-		}
-	}
-
-	// gen-foo creates foo.txt, and will always fail it's status check.
-	NewTaskTest(t,
-		WithName("run gen-foo 1 silent"),
-		WithExecutorOptions(
-			task.ExecutorWithDir(dir),
-			task.ExecutorWithSilent(true),
-		),
-		WithTask("gen-foo"),
-	)
-	// gen-foo creates bar.txt, and will pass its status-check the 3. time it
-	// is run. It creates bar.txt, but also lists it as its source. So, the checksum
-	// for the file won't match before after the second run as we the file
-	// only exists after the first run.
-	NewTaskTest(t,
-		WithName("run gen-bar 1 silent"),
-		WithExecutorOptions(
-			task.ExecutorWithDir(dir),
-			task.ExecutorWithSilent(true),
-		),
-		WithTask("gen-bar"),
-	)
-	// gen-silent-baz is marked as being silent, and should only produce output
-	// if e.Verbose is set to true.
-	NewTaskTest(t,
-		WithName("run gen-baz silent"),
-		WithExecutorOptions(
-			task.ExecutorWithDir(dir),
-			task.ExecutorWithSilent(true),
-		),
-		WithTask("gen-silent-baz"),
-	)
-
-	for _, f := range files {
-		if _, err := os.Stat(filepathext.SmartJoin(dir, f)); err != nil {
-			t.Errorf("File should exist: %v", err)
-		}
-	}
-
-	// Run gen-bar a second time to produce a checksum file that matches bar.txt
-	NewTaskTest(t,
-		WithName("run gen-bar 2 silent"),
-		WithExecutorOptions(
-			task.ExecutorWithDir(dir),
-			task.ExecutorWithSilent(true),
-		),
-		WithTask("gen-bar"),
-	)
-	// Run gen-bar a third time, to make sure we've triggered the status check.
-	NewTaskTest(t,
-		WithName("run gen-bar 3 silent"),
-		WithExecutorOptions(
-			task.ExecutorWithDir(dir),
-			task.ExecutorWithSilent(true),
-		),
-		WithTask("gen-bar"),
-	)
-
-	// Now, let's remove source file, and run the task again to to prepare
-	// for the next test.
-	err := os.Remove(filepathext.SmartJoin(dir, "bar.txt"))
-	require.NoError(t, err)
-	NewTaskTest(t,
-		WithName("run gen-bar 4 silent"),
-		WithExecutorOptions(
-			task.ExecutorWithDir(dir),
-			task.ExecutorWithSilent(true),
-		),
-		WithTask("gen-bar"),
-	)
-	// all: not up-to-date
-	NewTaskTest(t,
-		WithName("run gen-foo 2"),
-		WithExecutorOptions(
-			task.ExecutorWithDir(dir),
-		),
-		WithTask("gen-foo"),
-	)
-	// status: not up-to-date
-	NewTaskTest(t,
-		WithName("run gen-foo 3"),
-		WithExecutorOptions(
-			task.ExecutorWithDir(dir),
-		),
-		WithTask("gen-foo"),
-	)
-	// sources: not up-to-date
-	NewTaskTest(t,
-		WithName("run gen-bar 5"),
-		WithExecutorOptions(
-			task.ExecutorWithDir(dir),
-		),
-		WithTask("gen-bar"),
-	)
-	// all: up-to-date
-	NewTaskTest(t,
-		WithName("run gen-bar 6"),
-		WithExecutorOptions(
-			task.ExecutorWithDir(dir),
-		),
-		WithTask("gen-bar"),
-	)
-	// sources: not up-to-date, no output produced.
-	NewTaskTest(t,
-		WithName("run gen-baz 2"),
-		WithExecutorOptions(
-			task.ExecutorWithDir(dir),
-		),
-		WithTask("gen-silent-baz"),
-	)
-	// up-to-date, no output produced
-	NewTaskTest(t,
-		WithName("run gen-baz 3"),
-		WithExecutorOptions(
-			task.ExecutorWithDir(dir),
-		),
-		WithTask("gen-silent-baz"),
-	)
-	// up-to-date, output produced due to Verbose mode.
-	NewTaskTest(t,
-		WithName("run gen-baz 4 verbose"),
-		WithExecutorOptions(
-			task.ExecutorWithDir(dir),
-			task.ExecutorWithVerbose(true),
-		),
-		WithTask("gen-silent-baz"),
-		WithPostProcessFn(PPRemoveAbsolutePaths),
-	)
-}
-
-func TestPrecondition(t *testing.T) {
-	t.Parallel()
-	const dir = "testdata/precondition"
-	NewTaskTest(t,
-		WithName("a precondition has been met"),
-		WithExecutorOptions(
-			task.ExecutorWithDir(dir),
-		),
-		WithTask("foo"),
-	)
-	NewTaskTest(t,
-		WithName("a precondition was not met"),
-		WithExecutorOptions(
-			task.ExecutorWithDir(dir),
-		),
-		WithTask("impossible"),
-		WithRunError(),
-	)
-	NewTaskTest(t,
-		WithName("precondition in dependency fails the task"),
-		WithExecutorOptions(
-			task.ExecutorWithDir(dir),
-		),
-		WithTask("depends_on_impossible"),
-		WithRunError(),
-	)
-	NewTaskTest(t,
-		WithName("precondition in cmd fails the task"),
-		WithExecutorOptions(
-			task.ExecutorWithDir(dir),
-		),
-		WithTask("executes_failing_task_as_cmd"),
-		WithRunError(),
-	)
 }
 
 func TestGenerates(t *testing.T) {
@@ -914,250 +438,6 @@ func TestStatusChecksum(t *testing.T) { // nolint:paralleltest // cannot run in 
 			assert.Equal(t, time, s.ModTime())
 		})
 	}
-}
-
-func TestAlias(t *testing.T) {
-	t.Parallel()
-
-	NewTaskTest(t,
-		WithName("alias"),
-		WithExecutorOptions(
-			task.ExecutorWithDir("testdata/alias"),
-		),
-		WithTask("f"),
-	)
-
-	NewTaskTest(t,
-		WithName("duplicate alias"),
-		WithExecutorOptions(
-			task.ExecutorWithDir("testdata/alias"),
-		),
-		WithTask("x"),
-		WithRunError(),
-	)
-
-	NewTaskTest(t,
-		WithName("alias summary"),
-		WithExecutorOptions(
-			task.ExecutorWithDir("testdata/alias"),
-			task.ExecutorWithSummary(true),
-		),
-		WithTask("f"),
-	)
-}
-
-func TestLabel(t *testing.T) {
-	t.Parallel()
-
-	NewTaskTest(t,
-		WithName("up to date"),
-		WithExecutorOptions(
-			task.ExecutorWithDir("testdata/label_uptodate"),
-		),
-		WithTask("foo"),
-	)
-
-	NewTaskTest(t,
-		WithName("summary"),
-		WithExecutorOptions(
-			task.ExecutorWithDir("testdata/label_summary"),
-			task.ExecutorWithSummary(true),
-		),
-		WithTask("foo"),
-	)
-
-	NewTaskTest(t,
-		WithName("status"),
-		WithExecutorOptions(
-			task.ExecutorWithDir("testdata/label_status"),
-		),
-		WithTask("foo"),
-		WithStatusError(),
-	)
-
-	NewTaskTest(t,
-		WithName("var"),
-		WithExecutorOptions(
-			task.ExecutorWithDir("testdata/label_var"),
-		),
-		WithTask("foo"),
-	)
-
-	NewTaskTest(t,
-		WithName("label in summary"),
-		WithExecutorOptions(
-			task.ExecutorWithDir("testdata/label_summary"),
-		),
-		WithTask("foo"),
-	)
-}
-
-func TestPromptInSummary(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name      string
-		input     string
-		wantError bool
-	}{
-		{"test short approval", "y\n", false},
-		{"test long approval", "yes\n", false},
-		{"test uppercase approval", "Y\n", false},
-		{"test stops task", "n\n", true},
-		{"test junk value stops task", "foobar\n", true},
-		{"test Enter stops task", "\n", true},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			t.Parallel()
-
-			opts := []TaskTestOption{
-				WithName(test.name),
-				WithExecutorOptions(
-					task.ExecutorWithDir("testdata/prompt"),
-					task.ExecutorWithAssumeTerm(true),
-				),
-				WithTask("foo"),
-				WithInput(test.input),
-			}
-			if test.wantError {
-				opts = append(opts, WithRunError())
-			}
-			NewTaskTest(t, opts...)
-		})
-	}
-}
-
-func TestPromptWithIndirectTask(t *testing.T) {
-	t.Parallel()
-
-	NewTaskTest(t,
-		WithExecutorOptions(
-			task.ExecutorWithDir("testdata/prompt"),
-			task.ExecutorWithAssumeTerm(true),
-		),
-		WithTask("bar"),
-		WithInput("y\n"),
-	)
-}
-
-func TestPromptAssumeYes(t *testing.T) {
-	t.Parallel()
-
-	NewTaskTest(t,
-		WithName("--yes flag should skip prompt"),
-		WithExecutorOptions(
-			task.ExecutorWithDir("testdata/prompt"),
-			task.ExecutorWithAssumeTerm(true),
-			task.ExecutorWithAssumeYes(true),
-		),
-		WithTask("foo"),
-		WithInput("\n"),
-	)
-
-	NewTaskTest(t,
-		WithName("task should raise errors.TaskCancelledError"),
-		WithExecutorOptions(
-			task.ExecutorWithDir("testdata/prompt"),
-			task.ExecutorWithAssumeTerm(true),
-		),
-		WithTask("foo"),
-		WithInput("\n"),
-		WithRunError(),
-	)
-}
-
-func TestNoLabelInList(t *testing.T) {
-	t.Parallel()
-
-	const dir = "testdata/label_list"
-
-	var buff bytes.Buffer
-	e := task.NewExecutor(
-		task.ExecutorWithDir(dir),
-		task.ExecutorWithStdout(&buff),
-		task.ExecutorWithStderr(&buff),
-	)
-	require.NoError(t, e.Setup())
-	if _, err := e.ListTasks(task.ListOptions{ListOnlyTasksWithDescriptions: true}); err != nil {
-		t.Error(err)
-	}
-	assert.Contains(t, buff.String(), "foo")
-}
-
-// task -al case 1: listAll list all tasks
-func TestListAllShowsNoDesc(t *testing.T) {
-	t.Parallel()
-
-	const dir = "testdata/list_mixed_desc"
-
-	var buff bytes.Buffer
-	e := task.NewExecutor(
-		task.ExecutorWithDir(dir),
-		task.ExecutorWithStdout(&buff),
-		task.ExecutorWithStderr(&buff),
-	)
-	require.NoError(t, e.Setup())
-
-	var title string
-	if _, err := e.ListTasks(task.ListOptions{ListAllTasks: true}); err != nil {
-		t.Error(err)
-	}
-	for _, title = range []string{
-		"foo",
-		"voo",
-		"doo",
-	} {
-		assert.Contains(t, buff.String(), title)
-	}
-}
-
-// task -al case 2: !listAll list some tasks (only those with desc)
-func TestListCanListDescOnly(t *testing.T) {
-	t.Parallel()
-
-	const dir = "testdata/list_mixed_desc"
-
-	var buff bytes.Buffer
-	e := task.NewExecutor(
-		task.ExecutorWithDir(dir),
-		task.ExecutorWithStdout(&buff),
-		task.ExecutorWithStderr(&buff),
-	)
-	require.NoError(t, e.Setup())
-	if _, err := e.ListTasks(task.ListOptions{ListOnlyTasksWithDescriptions: true}); err != nil {
-		t.Error(err)
-	}
-
-	var title string
-	assert.Contains(t, buff.String(), "foo")
-	for _, title = range []string{
-		"voo",
-		"doo",
-	} {
-		assert.NotContains(t, buff.String(), title)
-	}
-}
-
-func TestListDescInterpolation(t *testing.T) {
-	t.Parallel()
-
-	const dir = "testdata/list_desc_interpolation"
-
-	var buff bytes.Buffer
-	e := task.NewExecutor(
-		task.ExecutorWithDir(dir),
-		task.ExecutorWithStdout(&buff),
-		task.ExecutorWithStderr(&buff),
-	)
-	require.NoError(t, e.Setup())
-	if _, err := e.ListTasks(task.ListOptions{ListOnlyTasksWithDescriptions: true}); err != nil {
-		t.Error(err)
-	}
-
-	assert.Contains(t, buff.String(), "foo-var")
-	assert.Contains(t, buff.String(), "bar-var")
 }
 
 func TestStatusVariables(t *testing.T) {
